refactor: update project models
