+++ packages/client/src/pages/Project/components/table/EnhancedTableToolbar.tsx
@@ -3,7 +3,7 @@
  * @param props
  */
 
-import React from 'react';
+import React, { FC, useState } from 'react';
 import { alpha } from '@mui/material/styles';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
@@ -11,13 +11,18 @@ import IconButton from '@mui/material/IconButton';
 import Tooltip from '@mui/material/Tooltip';
 import DeleteIcon from '@mui/icons-material/Delete';
 import { FormDialog } from '@pages/Project/components/form/FormDialog';
+import AddIcon from '@mui/icons-material/Add';
 
 interface EnhancedTableToolbarProps {
   numSelected: number;
 }
 
-export const EnhancedTableToolbar = (props: EnhancedTableToolbarProps) => {
-  const { numSelected } = props;
+export const EnhancedTableToolbar: FC<EnhancedTableToolbarProps> = ({ numSelected }) => {
+  const [open, setOpen] = useState(false);
+
+  const handleClickOpen = () => {
+    setOpen(true);
+  };
 
   return (
     <Toolbar
@@ -45,7 +50,22 @@ export const EnhancedTableToolbar = (props: EnhancedTableToolbarProps) => {
           </IconButton>
         </Tooltip>
       ) : (
-        <FormDialog type="add" />
+        <>
+          <IconButton
+            onClick={() => {
+              handleClickOpen();
+            }}
+            sx={{
+              '&:hover': {
+                backgroundColor: '#1565c0',
+                color: 'white'
+              }
+            }}
+          >
+            <AddIcon />
+          </IconButton>
+          <FormDialog open={open} setOpen={setOpen} title="Add Project" />
+        </>
       )}
     </Toolbar>
   );
